CORE-8608 Move CommandLineOrderingView into view package
