+++ de-lib/src/main/java/org/iplantc/de/apps/integration/client/gin/AppEditorGinModule.java
@@ -1,7 +1,7 @@
 package org.iplantc.de.apps.integration.client.gin;
 
 import org.iplantc.de.apps.integration.client.dialogs.CommandLineOrderingDialog;
-import org.iplantc.de.apps.integration.client.dialogs.CommandLineOrderingView;
+import org.iplantc.de.apps.integration.client.view.CommandLineOrderingView;
 import org.iplantc.de.apps.integration.client.gin.factory.DeployedComponentListingViewFactory;
 import org.iplantc.de.apps.integration.client.presenter.AppsEditorPresenterImpl;
 import org.iplantc.de.apps.integration.client.presenter.DeployedComponentPresenterImpl;
