+++ de-lib/src/main/java/org/iplantc/de/apps/integration/client/view/CommandLineOrderingView.java
@@ -1,7 +1,6 @@
-package org.iplantc.de.apps.integration.client.dialogs;
+package org.iplantc.de.apps.integration.client.view;
 
 import org.iplantc.de.apps.integration.client.model.ArgumentProperties;
-import org.iplantc.de.apps.integration.client.view.AppsEditorView;
 import org.iplantc.de.apps.widgets.client.view.editors.style.AppTemplateWizardAppearance;
 import org.iplantc.de.client.models.apps.integration.Argument;
