+++ de-lib/src/main/java/org/iplantc/de/apps/integration/client/view/CommandLineOrderingView.ui.xml
